feat: modified the critval function to be more general
