+++ NAMESPACE
@@ -22,6 +22,7 @@ export(calculate_semantic_sim_onto)
 export(calculate_tom)
 export(calculate_tom_from_exp)
 export(calculate_wang_sim)
+export(calculate_wang_sim_onto)
 export(change_gene_identifier)
 export(community_detection)
 export(contrastive_pca_processing)
@@ -99,6 +100,7 @@ export(rs_cor_upper_triangle)
 export(rs_cov2cor)
 export(rs_covariance)
 export(rs_create_random_aucs)
+export(rs_critval)
 export(rs_dense_to_upper_triangle)
 export(rs_differential_cor)
 export(rs_fast_auc)
@@ -114,7 +116,6 @@ export(rs_hypergeom_test_list)
 export(rs_ica_iters)
 export(rs_ica_iters_cv)
 export(rs_onto_semantic_sim)
-export(rs_onto_semantic_sim_critval)
 export(rs_onto_sim_wang)
 export(rs_ot_harmonic_sum)
 export(rs_phyper)
