+++ man/calculate_wang_sim_onto.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/methods_ontology.R
+\name{calculate_wang_sim_onto}
+\alias{calculate_wang_sim_onto}
+\title{Calculate the Wang similarity for an ontology.}
+\usage{
+calculate_wang_sim_onto(object, w = 0.8, .verbose = TRUE)
+}
+\arguments{
+\item{object}{\verb{ontology class}. See \code{\link[=ontology]{ontology()}}.}
+
+\item{w}{Float. The value assigned to the respective edge. Defaults to \code{0.8}.}
+
+\item{.verbose}{Boolean. Controls the verbosity of the function.}
+}
+\value{
+The class with added semantic similarities to the properties.
+}
+\description{
+This function calculates the Wang similarity for the whole
+ontology and adds it to the class. The current implementation just allows for
+a single w for the relationships.
+}
