+++ man/rs_critval.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R
+\name{rs_critval}
+\alias{rs_critval}
+\title{Calculate the critical value}
+\usage{
+rs_critval(values, iters, alpha, seed)
+}
+\arguments{
+\item{values}{Numeric vector. The full data set for which to calculate the
+critical value.}
+
+\item{iters}{Integer. Number of random permutations to use.}
+
+\item{alpha}{Float. The alpha value. For example, 0.001 would mean that the
+critical value is smaller than 0.1 percentile of the random permutations.}
+
+\item{seed}{Integer. For reproducibility purposes}
+}
+\value{
+The critical value for the given parameters.
+}
+\description{
+This function calculates the critical value for a given set based on random
+permutations and a given alpha value.
+}
