--- man/rs_onto_semantic_sim_critval.Rd
@@ -1,50 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/extendr-wrappers.R
-\name{rs_onto_semantic_sim_critval}
-\alias{rs_onto_semantic_sim_critval}
-\title{Calculate the semantic similarity in an ontology}
-\usage{
-rs_onto_semantic_sim_critval(
-  terms,
-  sim_type,
-  alpha,
-  ancestor_list,
-  ic_list,
-  iters,
-  seed
-)
-}
-\arguments{
-\item{terms}{Vector of strings. The terms in the ontology you wish to screen.}
-
-\item{sim_type}{String. Must be one of \code{c("resnik", "lin", "combined")}.}
-
-\item{alpha}{Float. Must be between 0 to 1. The alpha parameter for calculating
-the critival value.}
-
-\item{ancestor_list}{R list with names being the term and the elements in the
-list the names of the ancestors.}
-
-\item{ic_list}{R list with the names being the term and the elements the
-information content of this given term. Needs to be a single float!}
-
-\item{iters}{Integer. Number of random iterations to use to estimate the
-critical value.}
-
-\item{seed}{Integer. Random seed for reproducibility purposes.}
-}
-\value{
-A list with:
-\itemize{
-\item term1 - term 1 names.
-\item term1 - term 2 names.
-\item filtered_sim - the specified (filtered) similarity between the two
-terms.
-\item critval - the critical value for the given alpha.
-}
-}
-\description{
-This function calculates the specified semantic similarity and
-returns the full vector (only calculating the upper triangle) for the given
-similarity.
-}
